Fix bug: Change Position from i32 to f64

+++ query/src/doc/types/mod.rs
@@ -9,8 +9,8 @@ pub mod sort;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32
+    pub x: f64,
+    pub y: f64
 }
 
 #[derive(Debug, Clone, Hash, Eq, PartialEq, Serialize, Deserialize)]
